reformatting and added delete functionality
