+++ src/app/(ui)/(manager)/manage/people/components/EditPeopleCard.tsx
@@ -1,24 +1,21 @@
-"use client";
-
 import { useEffect, useRef, useState } from "react";
-import DateSelector from "@/components/DateSelector";
 import { Tables } from "@/lib/supabase.types";
+import DateSelector from "@/components/DateSelector";
 import { formatDate } from "@/utils";
-
-interface EditPeopleGridProps {
-  people: Tables<"users">[];
-  refreshPeople: () => Promise<void>;
-  filter: {
-    name: string;
-  };
-}
 
 interface EditPeopleCardProps {
   person: Tables<"users">;
   refreshPeople: () => Promise<void>;
+  onClick: () => void;
+  isSelected: boolean;
 }
 
-function EditPeopleCard({ person, refreshPeople }: EditPeopleCardProps) {
+export default function EditPeopleCard({
+  person,
+  refreshPeople,
+  onClick,
+  isSelected
+}: EditPeopleCardProps) {
   const [dob, setDob] = useState<Date | null>(
     person.dob ? new Date(person.dob) : null
   );
@@ -41,16 +38,29 @@ function EditPeopleCard({ person, refreshPeople }: EditPeopleCardProps) {
     };
   }, []);
 
+  let background = "bg-pedals-stroke";
+  if (isEditing) {
+    background = "bg-pedals-lightgrey";
+  } else if (isSelected) {
+    background = "bg-pedals-grey";
+  }
+
   return (
     <div
       ref={ref}
-      className={`edit-card flex w-full justify-between border-y-[1px] border-y-pedals-darkgrey px-20 py-4 ${isEditing ? "bg-pedals-lightgrey" : "bg-pedals-stroke"}`}
+      className={`edit-card flex w-full justify-between border-y-[1px] border-y-pedals-darkgrey px-20 py-4 ${background}`}
+      onClick={() => {
+        if (!isEditing) {
+          onClick();
+        }
+      }}
+      role="presentation"
     >
       <div className="flex items-center justify-start">
         <h3 className="w-80">{person.name ?? "Error"}</h3>
         <h2 className="w-80">{person.username ?? "Error"}</h2>
         {isEditing ? (
-          <div className="w-80 pr-4">
+          <div className="-ml-3 w-80 pr-4">
             <DateSelector selected={dob} onChange={(d) => setDob(d)} />
           </div>
         ) : (
@@ -69,46 +79,13 @@ function EditPeopleCard({ person, refreshPeople }: EditPeopleCardProps) {
       </div>
       <button
         type="button"
-        className="!rounded-[30px] !px-12 uppercase"
+        className={`!rounded-[30px] !px-12 uppercase ${isEditing && "hover:!bg-pedals-grey"}`}
         onClick={() => {
           setIsEditing(!isEditing);
         }}
       >
         {isEditing ? "Done" : "Edit"}
       </button>
-    </div>
-  );
-}
-
-export default function EditPeopleGrid({
-  people,
-  refreshPeople,
-  filter
-}: EditPeopleGridProps) {
-  return (
-    <div className="flex h-full flex-col overflow-y-auto">
-      <div className="sticky flex items-center justify-start px-20 py-2">
-        <p className="w-80">Name</p>
-        <p className="w-80">Username</p>
-        <p className="w-80">Date of Birth</p>
-        <p className="w-80">Last Seen</p>
-        <p className="ml-[340px]">Total</p>
-      </div>
-      <div className="overflow-y-scroll">
-        {people
-          ?.filter((person) =>
-            filter.name
-              ? person.name?.toLowerCase().includes(filter.name.toLowerCase())
-              : true
-          )
-          ?.map((person) => (
-            <EditPeopleCard
-              key={person.id}
-              person={person}
-              refreshPeople={refreshPeople}
-            />
-          ))}
-      </div>
     </div>
   );
 }
